backend/requirements.txt
# ===============================
# FastAPI Server
# ===============================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ===============================
# Database
# ===============================
SQLAlchemy==2.0.25
pymysql==1.1.0
mysql-connector-python==8.2.0

# ===============================
# AI / LLM
# ===============================
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# ===============================
# Google / Groq
# ===============================
google-generativeai==0.3.2
groq==0.4.2

# ===============================
# Vector & Math
# ===============================
numpy==1.24.3
scikit-learn==1.3.2

# ===============================
# Web Scraping
# ===============================
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ===============================
# Caching
# ===============================
redis==5.0.1
hiredis==2.3.2

# ===============================
# Cloud & Email
# ===============================
boto3==1.34.34
resend==0.8.0

# ===============================
# Utilities
# ===============================
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
loguru==0.7.2

# ===============================
# HTTP
# ===============================
httpx==0.26.0

# ===============================
# Testing
# ===============================
pytest==7.4.4
